+++ dashboard/src/components/form/FormInput/FormInput.tsx
@@ -1,3 +1,6 @@
+import getTransformedFieldProps, {
+  type Transformer,
+} from '@/components/form/utils/getTransformedFieldProps';
 import {
   FormControl,
   FormDescription,
@@ -9,15 +12,9 @@ import {
 import { Input, type InputProps } from '@/components/ui/v3/input';
 import { InfoTooltip } from '@/features/orgs/projects/common/components/InfoTooltip';
 import { cn, isNotEmptyValue } from '@/lib/utils';
-import {
-  type ChangeEvent,
-  type ForwardedRef,
-  forwardRef,
-  type ReactNode,
-} from 'react';
+import { type ForwardedRef, forwardRef, type ReactNode } from 'react';
 import type {
   Control,
-  ControllerRenderProps,
   FieldPath,
   FieldValues,
   PathValue,
@@ -36,9 +33,11 @@ interface FormInputProps<
   label: ReactNode;
   placeholder?: string;
   className?: string;
+  containerClassName?: string;
   type?: string;
   inline?: boolean;
   helperText?: string | null;
+  transform?: Transformer;
   transformValue?: (
     value: PathValue<TFieldValues, TName>,
   ) => PathValue<TFieldValues, TName>;
@@ -57,51 +56,31 @@ function InnerFormInput<
     label,
     placeholder,
     className = '',
+    containerClassName = '',
     type = 'text',
     inline,
     helperText,
-    transformValue,
     disabled,
     autoComplete,
     infoTooltip,
+    transform,
   }: FormInputProps<TFieldValues, TName>,
-  ref: ForwardedRef<HTMLInputElement>,
+  ref?: ForwardedRef<HTMLInputElement>,
 ) {
-  function getOnChangeHandlerAndValue(
-    field: ControllerRenderProps<TFieldValues, TName>,
-  ): [
-    PathValue<TFieldValues, TName>,
-    (e: ChangeEvent<HTMLInputElement>) => void,
-  ] {
-    const { onChange, value } = field;
-
-    function handleOnChange(event: ChangeEvent<HTMLInputElement>) {
-      let transformedValue = event.target.value;
-      if (isNotEmptyValue(transformValue)) {
-        transformedValue = transformValue(
-          event.target.value as PathValue<TFieldValues, TName>,
-        );
-      }
-      onChange(transformedValue);
-    }
-
-    const transformedValue = isNotEmptyValue(transformValue)
-      ? transformValue(value)
-      : value;
-
-    return [transformedValue, handleOnChange];
-  }
   return (
     <FormField
       control={control}
       name={name}
       render={({ field }) => {
-        const { onChange, value, ...fieldProps } = field;
-
-        const [tValue, handleOnChange] = getOnChangeHandlerAndValue(field);
+        const fieldProps = isNotEmptyValue(transform)
+          ? getTransformedFieldProps(field, transform)
+          : field;
         return (
           <FormItem
-            className={cn({ 'flex w-full items-center gap-4 py-3': inline })}
+            className={cn(
+              { 'flex w-full items-center gap-4 py-3': inline },
+              containerClassName,
+            )}
           >
             {infoTooltip ? (
               <div className="flex flex-row items-center gap-2">
@@ -133,8 +112,6 @@ function InnerFormInput<
                 <Input
                   type={type}
                   placeholder={placeholder}
-                  onChange={handleOnChange}
-                  value={tValue}
                   disabled={disabled}
                   autoComplete={autoComplete}
                   {...fieldProps}