+++ dashboard/src/features/orgs/projects/database/dataGrid/components/DataBrowserSidebar/DataBrowserSidebar.tsx
@@ -14,6 +14,7 @@ import {
 } from '@/components/ui/v3/select';
 import { Spinner } from '@/components/ui/v3/spinner';
 import { useIsPlatform } from '@/features/orgs/projects/common/hooks/useIsPlatform';
+import { EditTableSettingsForm } from '@/features/orgs/projects/database/dataGrid/components/EditTableSettingsForm';
 import { useDatabaseQuery } from '@/features/orgs/projects/database/dataGrid/hooks/useDatabaseQuery';
 import { useDeleteTableWithToastMutation } from '@/features/orgs/projects/database/dataGrid/hooks/useDeleteTableMutation';
 import { isSchemaLocked } from '@/features/orgs/projects/database/dataGrid/utils/schemaHelpers';
@@ -267,6 +268,34 @@ function DataBrowserSidebarContent({
     });
   }
 
+  function handleEditSettingsClick(
+    schema: string,
+    table: string,
+    disabled?: boolean,
+  ) {
+    openDrawer({
+      title: (
+        <span className="inline-grid grid-flow-col items-center gap-2">
+          {disabled ? 'View settings for' : 'Edit settings for'}
+          <InlineCode className="!text-sm+ font-normal">{table}</InlineCode>
+          table
+        </span>
+      ),
+      component: (
+        <EditTableSettingsForm
+          disabled={disabled}
+          schema={schema}
+          tableName={table}
+        />
+      ),
+      props: {
+        PaperProps: {
+          className: 'overflow-hidden ',
+        },
+      },
+    });
+  }
+
   return (
     <div className="flex h-full flex-col justify-between">
       <div className="box flex flex-col px-2">
@@ -385,6 +414,13 @@ function DataBrowserSidebarContent({
                               true,
                             )
                           }
+                          onViewSettings={() =>
+                            handleEditSettingsClick(
+                              table.table_schema,
+                              table.table_name,
+                              true,
+                            )
+                          }
                           onEditTable={() =>
                             openDrawer({
                               title: 'Edit Table',
@@ -408,6 +444,13 @@ function DataBrowserSidebarContent({
                               table.table_name,
                             )
                           }
+                          onEditSettings={() => {
+                            handleEditSettingsClick(
+                              table.table_schema,
+                              table.table_name,
+                              false,
+                            );
+                          }}
                           onDelete={() =>
                             handleDeleteTableClick(
                               table.table_schema,