+++ dashboard/src/features/orgs/projects/database/dataGrid/hooks/useCreateTableMutation/prepareCreateTableQuery.test.ts
@@ -224,8 +224,8 @@ describe('prepareCreateTableQuery', () => {
         {
           name: 'name',
           type: {
-            value: 'varchar(10)' as any,
-            label: 'varchar(10)',
+            value: 'character varying(10)' as any,
+            label: 'character varying(10)',
           },
         },
       ],
@@ -240,7 +240,7 @@ describe('prepareCreateTableQuery', () => {
 
     expect(transaction).toHaveLength(1);
     expect(transaction[0].args.sql).toBe(
-      'CREATE TABLE public.test_table (id uuid NOT NULL, name varchar(10) NOT NULL);',
+      'CREATE TABLE public.test_table (id uuid NOT NULL, name character varying(10) NOT NULL);',
     );
   });