+++ dashboard/src/features/orgs/projects/database/dataGrid/hooks/useTableIsEnumQuery/index.ts
@@ -1,0 +1,1 @@
+export { default as useTableIsEnumQuery } from './useTableIsEnumQuery';